requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.6
logzero==1.3.0
python-magic==0.4.13
qrcode==5.3
requests==2.18.4
six==1.10.0
urllib3==1.22
